requirements.txt
Django==4.2.12
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-filter==24.3
django-cors-headers==4.6.0
Pillow>=12.1.1
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.7
django-celery-beat==2.5.0